apps/api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
psycopg2
python-dotenv
requests
pyjwt
cryptography
stripe